requirements.txt
datasets==3.5.1
numpy==2.1.2
peft==0.17.0
Pillow==11.0.0
polars==1.35.2
PyYAML==6.0.2
qwen_vl_utils==0.0.11
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3